fix(test): consolidate ScyllaDB tests into tier2 to eliminate container leak
The static OnceCell in canon-event-store-cassandra leaked its ScyllaDB
container (Rust does not run Drop on statics), preventing the tier2
tests from starting a second container on resource-constrained CI.

Fix: move all 3 CassandraEventStore integration tests (append_and_load,
optimistic_concurrency_conflict, load_from_version) into
canon-test/tests/tier2_e2e.rs where they share the single ScyllaDB
container with the other tier2 Cassandra tests. This eliminates the
need for two ScyllaDB containers across the test suite.

- Remove test module and testcontainers dev-deps from
  canon-event-store-cassandra
- Add tier2_cassandra_append_and_load,
  tier2_cassandra_optimistic_concurrency_conflict,
  tier2_cassandra_load_from_version to tier2_e2e.rs
- All tests share the existing OnceLock<OnceCell<ScyllaContainer>>
