+++ canon-event-store-cassandra/src/lib.rs
@@ -387,184 +387,10 @@ fn row_to_envelope(row: EventRow) -> Result<EventEnvelope, CassandraEventStoreEr
 }
 
 // ── Tests ───────────────────────────────────────────────────────────────────
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use testcontainers::runners::AsyncRunner;
-    use testcontainers::ContainerAsync;
-    use testcontainers::ImageExt;
-    use testcontainers_modules::scylladb::ScyllaDB;
-    use tokio::sync::OnceCell;
-
-    // ── Shared container ────────────────────────────────────────────────
-
-    /// Single ScyllaDB container shared across all tests in this module.
-    ///
-    /// The `OnceCell` holds both the container handle (to keep it alive for
-    /// the lifetime of the test binary) and the connection URI. The first
-    /// test to call `shared_scylla_uri()` starts the container; subsequent
-    /// tests reuse it. Each test creates its own `CassandraEventStore`
-    /// session (cheap) and uses a unique `AggregateId`, so there is no
-    /// data collision.
-    ///
-    /// Note: Rust does not run destructors for statics, so the Docker
-    /// container will be orphaned when the test process exits. This is the
-    /// standard tradeoff for shared testcontainers — Docker's cleanup
-    /// policies or the next test run will handle removal.
-    static SCYLLA: OnceCell<(ContainerAsync<ScyllaDB>, String)> = OnceCell::const_new();
-
-    /// Returns the connection URI for the shared ScyllaDB container,
-    /// starting it on first call.
-    async fn shared_scylla_uri() -> &'static str {
-        let (_, uri) = SCYLLA
-            .get_or_init(|| async {
-                let container = ScyllaDB::default()
-                    .with_startup_timeout(std::time::Duration::from_secs(120))
-                    .start()
-                    .await
-                    .expect("Failed to start ScyllaDB container");
-
-                let host = container.get_host().await.expect("host");
-                let port = container.get_host_port_ipv4(9042_u16).await.expect("port");
-                let uri = format!("{host}:{port}");
-
-                // ScyllaDB may need a moment after the ready condition before it
-                // can accept CQL queries. Retry the initial connection.
-                let session = {
-                    let mut attempts = 0u32;
-                    loop {
-                        match SessionBuilder::new().known_node(&uri).build().await {
-                            Ok(s) => break s,
-                            Err(e) => {
-                                attempts += 1;
-                                if attempts > 30 {
-                                    panic!("Failed to connect to ScyllaDB after 30 attempts: {e}");
-                                }
-                                tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-                            }
-                        }
-                    }
-                };
-
-                // Create keyspace and table
-                session
-                    .query_unpaged(
-                        "CREATE KEYSPACE IF NOT EXISTS canon \
-                         WITH replication = {'class': 'SimpleStrategy', 'replication_factor': 1}",
-                        &[],
-                    )
-                    .await
-                    .expect("create keyspace");
-
-                session
-                    .query_unpaged(
-                        "CREATE TABLE IF NOT EXISTS canon.events ( \
-                         aggregate_id UUID, version BIGINT, event_id UUID, event_type TEXT, \
-                         event_version INT, payload BLOB, correlation_id UUID, causation_id UUID, \
-                         created_at TIMESTAMP, \
-                         PRIMARY KEY (aggregate_id, version) \
-                         ) WITH CLUSTERING ORDER BY (version ASC)",
-                        &[],
-                    )
-                    .await
-                    .expect("create events table");
-
-                (container, uri)
-            })
-            .await;
-        uri.as_str()
-    }
-
-    /// Creates a `CassandraEventStore` connected to the shared ScyllaDB
-    /// container. Each test gets its own session (cheap to create).
-    async fn setup_store() -> CassandraEventStore {
-        let uri = shared_scylla_uri().await;
-        CassandraEventStore::new(uri)
-            .await
-            .expect("failed to create CassandraEventStore")
-    }
-
-    // ── Helpers ─────────────────────────────────────────────────────────
-
-    fn make_event(aggregate_id: &AggregateId) -> EventEnvelope {
-        EventEnvelope {
-            event_id: Uuid::new_v4(),
-            aggregate_id: aggregate_id.clone(),
-            version: Version::initial(),
-            event_type: "TestEvent".into(),
-            event_version: 1,
-            payload: Bytes::from_static(b"{}"),
-            correlation_id: Uuid::new_v4(),
-            causation_id: Uuid::new_v4(),
-            timestamp: Utc::now(),
-        }
-    }
-
-    // ── Integration tests ───────────────────────────────────────────────
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_append_and_load() {
-        let store = setup_store().await;
-        let agg_id = AggregateId::new();
-
-        let events = vec![make_event(&agg_id), make_event(&agg_id)];
-        store
-            .append(&agg_id, Version::initial(), events)
-            .await
-            .expect("append should succeed");
-
-        let loaded = store.load(&agg_id).await.expect("load should succeed");
-        assert_eq!(loaded.len(), 2);
-        assert_eq!(loaded[0].version.as_u64(), 1);
-        assert_eq!(loaded[1].version.as_u64(), 2);
-        assert_eq!(loaded[0].event_type, "TestEvent");
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_optimistic_concurrency_conflict() {
-        let store = setup_store().await;
-        let agg_id = AggregateId::new();
-
-        store
-            .append(&agg_id, Version::initial(), vec![make_event(&agg_id)])
-            .await
-            .expect("first append should succeed");
-
-        // Second append at same expected_version must fail
-        let result = store
-            .append(&agg_id, Version::initial(), vec![make_event(&agg_id)])
-            .await;
-
-        assert!(
-            matches!(result, Err(EventStoreError::VersionConflict { .. })),
-            "expected VersionConflict, got: {result:?}",
-        );
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_load_from_version() {
-        let store = setup_store().await;
-        let agg_id = AggregateId::new();
-
-        let events = vec![
-            make_event(&agg_id),
-            make_event(&agg_id),
-            make_event(&agg_id),
-        ];
-        store
-            .append(&agg_id, Version::initial(), events)
-            .await
-            .expect("append should succeed");
-
-        // Load from version 2 — should return events at version 2 and 3
-        let loaded = store
-            .load_from_version(&agg_id, Version::from_u64(2))
-            .await
-            .expect("load_from_version should succeed");
-
-        assert_eq!(loaded.len(), 2);
-        assert_eq!(loaded[0].version.as_u64(), 2);
-        assert_eq!(loaded[1].version.as_u64(), 3);
-    }
-}
+//
+// Integration tests for CassandraEventStore live in canon-test/tests/tier2_e2e.rs
+// alongside other testcontainers-based tests. This avoids starting a second
+// ScyllaDB container (statics don't run destructors, so a container held in a
+// static OnceCell leaks and competes for resources with the tier2 container on
+// CI). All three original tests (append_and_load, optimistic_concurrency_conflict,
+// load_from_version) have been migrated there.
